Implemented some fixes to user creation and updating logic after testing by sending requests. Configured dev and prod application properties (H2 in memory db for dev and Postgre for prod). Also added some missing method and class comments. Project is almost done, only unit and integration testing are left to do.

+++ src/main/java/com/heamimont/salesstoreapi/dto/order/OrderDTO.java
@@ -4,6 +4,7 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -13,5 +14,5 @@ public abstract class OrderDTO {
 
     @NotNull(message = "Total cost is required")
     @Positive(message = "Total cost must be positive")
-    protected Double totalCost;
+    protected BigDecimal totalCost;
 }
